+++ openbb_terminal/miscellaneous/library/trail_map.csv
@@ -187,7 +187,7 @@ crypto.ov.cpmarkets,openbb_terminal.cryptocurrency.overview.coinpaprika_model.ge
 crypto.ov.cpcontracts,openbb_terminal.cryptocurrency.overview.coinpaprika_model.get_contract_platform,openbb_terminal.cryptocurrency.overview.coinpaprika_view.display_contracts
 crypto.ov.cpexmarkets,openbb_terminal.cryptocurrency.overview.coinpaprika_model.get_exchanges_market,openbb_terminal.cryptocurrency.overview.coinpaprika_view.display_exchange_markets
 crypto.ov.cpglobal,openbb_terminal.cryptocurrency.overview.coinpaprika_model.get_global_market,openbb_terminal.cryptocurrency.overview.coinpaprika_view.display_global_market
-crypto.ov.list_of_coins,openbb_terminal.cryptocurrency.overview.coinpaprika_model.get_list_of_coins,
+crypto.ov.list_of_coins,openbb_terminal.cryptocurrency.cryptocurrency_helpers.get_list_of_coins,
 crypto.ov.cpexchanges,openbb_terminal.cryptocurrency.overview.coinpaprika_model.get_list_of_exchanges,openbb_terminal.cryptocurrency.overview.coinpaprika_view.display_all_exchanges
 crypto.ov.news,openbb_terminal.cryptocurrency.overview.cryptopanic_model.get_news,openbb_terminal.cryptocurrency.overview.cryptopanic_view.display_news
 crypto.ov.cr,openbb_terminal.cryptocurrency.overview.loanscan_model.get_rates,openbb_terminal.cryptocurrency.overview.loanscan_view.display_crypto_rates
